# openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Drammers.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/events": {
      "get": {
        "tags": [
          "AdminEvents"
        ],
        "parameters": [
          {
            "name": "includePast",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminEventSummaryResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminEventSummaryResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminEventSummaryResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminEvents"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/events/{id}": {
      "get": {
        "tags": [
          "AdminEvents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEventResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEventResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEventResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AdminEvents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AdminEvents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "highlight",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfEventSummaryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfEventSummaryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfEventSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{id}": {
      "get": {
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{id}/ical": {
      "get": {
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/calendar": { }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/calendar": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminEventResponse": {
        "required": [
          "id",
          "categoryId",
          "title",
          "summary",
          "description",
          "startAt",
          "endAt",
          "allDay",
          "locationName",
          "locationAddress",
          "latitude",
          "longitude",
          "isHighlight",
          "badgeText",
          "publication",
          "imageUrl",
          "attachments"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "allDay": {
            "type": "boolean"
          },
          "locationName": {
            "type": [
              "null",
              "string"
            ]
          },
          "locationAddress": {
            "type": [
              "null",
              "string"
            ]
          },
          "latitude": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "longitude": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "isHighlight": {
            "type": "boolean"
          },
          "badgeText": {
            "type": [
              "null",
              "string"
            ]
          },
          "publication": {
            "$ref": "#/components/schemas/PublicationResponse"
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttachmentResponse"
            }
          }
        }
      },
      "AdminEventSummaryResponse": {
        "required": [
          "id",
          "title",
          "startAt",
          "visibility",
          "status",
          "publishAt",
          "isHighlight"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "visibility": {
            "$ref": "#/components/schemas/ContentVisibility"
          },
          "status": {
            "$ref": "#/components/schemas/PublicationStatus"
          },
          "publishAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "isHighlight": {
            "type": "boolean"
          }
        }
      },
...
      "AttachmentResponse": {
        "required": [
          "id",
          "fileName",
          "contentType",
          "sizeBytes",
          "url"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "url": {
            "type": "string"
          }
        }
      },
...
      "ContentVisibility": {
        "enum": [
          "Public",
          "Members",
          "Restricted"
        ],
        "description": "Wie content mag zien (docs/07 §4). Beheren gaat via permissions, niet via zichtbaarheid."
      },
      "CreatedResponse": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "EventCategoryResponse": {
        "required": [
          "id",
          "code",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "EventDetailResponse": {
        "required": [
          "id",
          "title",
          "summary",
          "descriptionHtml",
          "startAt",
          "endAt",
          "allDay",
          "locationName",
          "locationAddress",
          "latitude",
          "longitude",
          "category",
          "isHighlight",
          "badgeText",
          "imageUrl",
          "attachments"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "descriptionHtml": {
            "type": [
              "null",
              "string"
            ]
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "allDay": {
            "type": "boolean"
          },
          "locationName": {
            "type": [
              "null",
              "string"
            ]
          },
          "locationAddress": {
            "type": [
              "null",
              "string"
            ]
          },
          "latitude": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "longitude": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "category": {
            "$ref": "#/components/schemas/EventCategoryResponse"
          },
          "isHighlight": {
            "type": "boolean"
          },
          "badgeText": {
            "type": [
              "null",
              "string"
            ]
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttachmentResponse"
            }
          }
        }
      },
      "EventRequest": {
        "required": [
          "categoryId",
          "title",
          "summary",
          "description",
          "startAt",
          "endAt",
          "allDay",
          "locationName",
          "locationAddress",
          "latitude",
          "longitude",
          "isHighlight",
          "badgeText",
          "publication"
        ],
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "allDay": {
            "type": "boolean"
          },
          "locationName": {
            "type": [
              "null",
              "string"
            ]
          },
          "locationAddress": {
            "type": [
              "null",
              "string"
            ]
          },
          "latitude": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "longitude": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "isHighlight": {
            "type": "boolean"
          },
          "badgeText": {
            "type": [
              "null",
              "string"
            ]
          },
          "publication": {
            "$ref": "#/components/schemas/PublicationRequest"
          }
        }
      },
      "EventSummaryResponse": {
        "required": [
          "id",
          "title",
          "summary",
          "startAt",
          "endAt",
          "allDay",
          "locationName",
          "category",
          "isHighlight",
          "badgeText",
          "imageUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "allDay": {
            "type": "boolean"
          },
          "locationName": {
            "type": [
              "null",
              "string"
            ]
          },
          "category": {
            "$ref": "#/components/schemas/EventCategoryResponse"
          },
          "isHighlight": {
            "type": "boolean"
          },
          "badgeText": {
            "type": [
              "null",
              "string"
            ]
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PagedResultOfEventSummaryResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventSummaryResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "Pagineringsconventie van de API (docs/05 §1): `?page=1&amp;pageSize=25`, maximaal 100."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PublicationRequest": {
        "required": [
          "visibility",
          "audienceRoles",
          "status",
          "publishAt"
        ],
        "type": "object",
        "properties": {
          "visibility": {
            "$ref": "#/components/schemas/ContentVisibility"
          },
          "audienceRoles": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "status": {
            "$ref": "#/components/schemas/PublicationStatus"
          },
          "publishAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "PublicationResponse": {
        "required": [
          "visibility",
          "audienceRoles",
          "status",
          "publishAt"
        ],
        "type": "object",
        "properties": {
          "visibility": {
            "$ref": "#/components/schemas/ContentVisibility"
          },
          "audienceRoles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "$ref": "#/components/schemas/PublicationStatus"
          },
          "publishAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "PublicationStatus": {
        "enum": [
          "Draft",
          "Scheduled",
          "Published",
          "Archived"
        ]
      },
...
    }
  },
...
}